Day 11 is a bad day
Got up in my head, built the entire thing as a list of points with values
after moving to part two I realized that was a huge mistake for this problem
and switched back to a 2d array.

By the time I managed to finish my brain was so fried that I wrote some
truely bad code for solving part 2.