Day 14 - Huge Fun
So, this one was great in my opinion, first one building up to the
second one. Second one seems like a very different problem, but
is really super similar.

I went down a small wrong path in trying to use the mask to generate
at the time that I was also trying to deal with the register, but
eventually realized that was a mistake, and instead generated all
possible masks, and then just used them all...worked really well.