Day23 Stolen
I did a "linkedlist" solution, that should have worked, but then
ended up taking over an hour to get to a reasonable number. At
which point I realized that I had been doing a linkedlist of
10 million, rather than 1 million.

By this time I had also implemented a MUCH better solution that
takes miliseconds to run the simulation. And after over an hour
of debugging that (while the above ran) found my bug in the
above, fixed it, and solved the problem...bad solution lives on
for posterity...