Day 24
I really enjoyed this day, I saw right away creating a cononical ID
for a tile was the answer, and so started working on a concept of
simplifying directions.

From there both parts 1 and 2 were reasonably straight forward,
where 1 was just counting the number of times we had a "cell", and
part 2 was just the same problem we've solved 2 other times this
year.